Moved examples and removed red lines on car details
