+++ package.json
@@ -11,8 +11,8 @@
   "dependencies": {
     "@expo/webpack-config": "^19.0.0",
     "@react-navigation/bottom-tabs": "^6.5.8",
-    "@react-navigation/native-stack": "^6.9.13",
     "@react-navigation/native": "^6.1.7",
+    "@react-navigation/native-stack": "^6.9.13",
     "expo": "~49.0.8",
     "expo-status-bar": "~1.6.0",
     "react": "18.2.0",
@@ -23,6 +23,7 @@
   "devDependencies": {
     "@babel/core": "^7.20.0",
     "@types/react": "~18.2.14",
+    "eslint-plugin-react-hooks": "^4.6.0",
     "typescript": "^5.1.3"
   },
   "private": true
